Add a Concourse pipeline for testing docker image builds
This pipeline builds a docker image in one job and consumes it in another, using
`passed` to get only image versions that came from the `build-the-image` job.
This was done as a test to make sure that no other versions got through to the
`get-the-image` job, even when newer images were pushed to the repo.